gmodule: fix warning reported by MSVC

+++ gmodule/gmodule-win32.c
@@ -168,7 +168,7 @@ static gchar*
 _g_module_build_path (const gchar *directory,
 		      const gchar *module_name)
 {
-  gint k;
+  gsize k;
 
   k = strlen (module_name);
